ver 0.4

+++ src/main/java/pl/database/Entity/IssLocation.java
@@ -60,7 +60,7 @@ public class IssLocation {
         this.timestamp = timestamp;
     }
 
-    public int getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
